+++ README.md
@@ -4,4 +4,4 @@
 Simple C++ classes for Object Relational Mapping
 
 ### Dependencies:
-`sudo apt-get install libmysqlcppconn-dev libboost-filesystem-dev libboost-program-options-dev`
+`sudo apt-get install libmysqlcppconn-dev`
