+++ tst/compile_test.sh
